+++ pod-operation/src/components/signal_light.rs
@@ -1,31 +1,31 @@
-// use tracing::debug;
+use tracing::debug;
 
-// use rppal::gpio::{Gpio, OutputPin};
+use rppal::gpio::{Gpio, OutputPin};
 
-// pub struct SignalLight {
-// 	pin: OutputPin,
-// }
+pub struct SignalLight {
+	pin: OutputPin,
+}
 
-// const PIN_SIGNAL_LIGHT: u8 = 21;
+const PIN_SIGNAL_LIGHT: u8 = 21;
 
-// impl SignalLight {
-// 	pub fn new() -> Self {
-// 		SignalLight {
-// 			pin: Gpio::new()
-// 				.unwrap()
-// 				.get(PIN_SIGNAL_LIGHT)
-// 				.unwrap()
-// 				.into_output(),
-// 		}
-// 	}
+impl SignalLight {
+	pub fn new() -> Self {
+		SignalLight {
+			pin: Gpio::new()
+				.unwrap()
+				.get(PIN_SIGNAL_LIGHT)
+				.unwrap()
+				.into_output(),
+		}
+	}
 
-// 	pub fn disable(&mut self) {
-// 		debug!("Disabling signal light.");
-// 		self.pin.set_low();
-// 	}
+	pub fn disable(&mut self) {
+		debug!("Disabling signal light.");
+		self.pin.set_low();
+	}
 
-// 	pub fn enable(&mut self) {
-// 		debug!("Enabling signal light.");
-// 		self.pin.set_high();
-// 	}
-// }
+	pub fn enable(&mut self) {
+		debug!("Enabling signal light.");
+		self.pin.set_high();
+	}
+}
